+++ examples/butler/main.go
@@ -30,8 +30,8 @@ func app() *warg.App {
 				),
 			),
 			section.ExistingCommand("version", warg.VersionCommand()),
+			section.ExistingFlag("--color", warg.ColorFlag()),
 		),
-		warg.AddColorFlag(),
 		warg.SkipValidation(),
 	)
 	return &app