+++ examples/fling/main.go
@@ -71,8 +71,8 @@ func app() *warg.App {
 			),
 			section.ExistingCommand("version", warg.VersionCommand()),
 			section.Footer("Homepage: https://github.com/bbkane/fling"),
+			section.ExistingFlag("--color", warg.ColorFlag()),
 		),
-		warg.AddColorFlag(),
 		warg.SkipValidation(),
 	)
 	return &app