+++ command/command.go
@@ -1,6 +1,7 @@
 package command
 
 import (
+	"context"
 	"errors"
 	"os"
 	"sort"
@@ -15,11 +16,16 @@ type PassedFlags map[string]interface{} // This can just stay a string for the c
 // Context holds everything a command needs.
 type Context struct {
 	AppName string
+
+	// Context to smuggle user-defined state (i.e., not flags) into an Action. I use this for mocks when testing
+	Context context.Context
+	Flags   PassedFlags
+
 	// Path passed either to a command or a section. Does not include executable name (os.Args[0])
 	Path   []string
-	Flags  PassedFlags
 	Stderr *os.File
 	Stdout *os.File
+
 	// Version of this app
 	Version string
 }
@@ -57,6 +63,7 @@ type CommandOpt func(*Command)
 // The name of a Command should probably be a verb - add , edit, run, ...
 // A Command should not be constructed directly. Use Command / New / ExistingCommand functions
 type Command struct {
+	// Action to run when command is invoked
 	Action Action
 
 	// Parsed Flags