+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Some versioning notes: warg's primary purpose is to serve my (@bbkane) CLIs, so
 it's likely to have breaking changes whenever I discover a better way to do
 something and remain < 1.0.0 forever.
 
-## Unreleased
+## v0.0.21
 
 ## Added